Daemon: avoid exceptions on premature shutdown
In case of error during startup, the setup is aborted and a
shutdown event is fired by the reactor.

When this happens during the Session setup phase, the blob_manager
may remain uninitialized, thus triggering errors in other routines.

Prevents those errors by checking for the blob_manager to be
initialized before accessing it in other setup routines.

Signed-off-by: Antonio Quartulli <antonio@mandelbit.com>